refactorize to version 0.1.x
