+++ src/command/import.ts
@@ -1,7 +1,8 @@
 /* eslint-disable no-mixed-spaces-and-tabs */
 import { CommandModule, Argv, Arguments } from 'yargs'
-import { Orm, Helper } from 'lambdaorm'
+import { Orm } from 'lambdaorm'
 import path from 'path'
+import { Manager } from '../manager'
 
 export class ImportCommand implements CommandModule {
 	command = 'import';
@@ -17,35 +18,39 @@ export class ImportCommand implements CommandModule {
 				alias: 'stage',
 				describe: 'Name of stage'
 			})
-			.option('s', {
-				alias: 'source',
-				describe: 'Source file to import.'
+			.option('d', {
+				alias: 'data',
+				describe: 'Data file to import.'
 			})
 	}
 
 	async handler (args: Arguments) {
 		const workspace = path.resolve(process.cwd(), args.workspace as string || '.')
 		const stageName = args.stage as string
-		const source = args.source as string
-		const orm = new Orm(workspace)
+		const data = args.data || {}
 
-		if (source === undefined) {
-			console.error('the source argument is required')
+		if (data === undefined) {
+			console.error('the data argument is required')
 			return
 		}
+		const orm = new Orm(workspace)
 
 		try {
 			const schema = await orm.schema.get(workspace)
-			const stage = orm.schema.stage.get(stageName)
 			await orm.init(schema)
-			// get content
-			const content = await Helper.readFile(source)
-			if (content === null) {
-				throw new Error(`source: ${source} not found or empty`)
-			}
-			// import data
-			const data = JSON.parse(content)
-			await orm.stage.import(stage.name).execute(data)
+			const stage = orm.schema.stage.get(stageName)
+			const manager = new Manager(orm)
+			// read Data
+			const _data = await manager.readData(data)
+
+			// // get content
+			// const content = await Helper.readFile(source)
+			// if (content === null) {
+			// throw new Error(`source: ${source} not found or empty`)
+			// }
+			// // import data
+			// const data = JSON.parse(content)
+			await orm.stage.import(stage.name).execute(_data)
 		} catch (error) {
 			console.error(`error: ${error}`)
 		} finally {
